benchmarks/synth/requirements.txt
suite2p
antspyx
itk-elastix
caiman
h5py
tifffile
numpy
scipy
pandas
opencv-python
